Adressed requested changes from PR25
