+++ views/connected.html
@@ -3,7 +3,6 @@
   <link rel="stylesheet" href="//dd04rofzygnm0.cloudfront.net/releases/master/865f04865d2448286626bac92c518a8f8ea8bafe/stylesheets/neue.css">
   <script type="text/javascript" charset="utf-8">
     (function() { 
-      console.log("Query",  "<%- query %>");
       window.original_query = "<%- query %>";
     })();
   </script>
